\name{gbm}
\alias{gbm}
\alias{NCHgbm450}
\alias{TCGAgbm27}
\docType{data}
\title{
DNA methylation of MGMT promoter region from Infinium HM-450k and HM-27K platforms
}
\description{
DNA methylation of MGMT promoter region from Infinium HM-450k and HM-27K platforms (table S3 and S4 described in details Bady et al. 2012).
}
\usage{data("gbm")}
\format{
  two data.frames corresponding to the tables S3 and S4 described in details Bady et al. (2012)
}
\details{
The results published here are in part based upon data generated by The Cancer TCGA Genome Atlas pilot project established by the NCI and NHGRI. Information about TCGA and the investigators and institutions who constitute the TCGA research network can be found at (http://cancergenome.nih.gov). The dbGaP accession number to the specific version of the TCGA data set is phs000178.v8.p7.
}
\source{
This two datasets correspond to the tables S3 and S4 from Bady et al. (2012).
}
\references{
Bady, P., D. Sciuscio, A.-C. Diserens, J. Bloch, M. J. van den Bent, C. Marosi, P.-Y. Dietrich, M. Weller, L. Mariani, F. L. Heppner, D. R. McDonald, D. Lacombe, R. Stupp, M. Delorenzi, and M. E. Hegi. 2012. MGMT methylation analysis of glioblastoma on the Infinium methylation BeadChip identifies two distinct CpG regions associated with gene silencing and outcome, yielding a prediction model for comparisons across datasets, tumor grades, and CIMP-status. Acta Neuropathologica 124:547-560.\cr
The Cancer Genome Atlas Consortium Comprehensive genomic characterization defines human glioblastoma genes and core pathways. Nature. 2008;455:1061-1068.\cr
Noushmehr, H., D. Weisenberger, K. Diefes, H. Phillips, K. Pujara, B. Berman, F. Pan, C. Pelloski, E. Sulman, and K. Bhat. 2010. Identification of a CpG island methylator phenotype that defines a distinct subgroup of glioma. Cancer Cell 17:510 - 522.\cr
}
\examples{
# table S3 (bady et al 2012)
data(NCHgbm450)
head(NCHgbm450)
# table S4 (bady et al 2012)
data(TCGAgbm27)
head(TCGAgbm27)
}
\keyword{datasets}
