\name{mgmtstp27-package}
\alias{mgmtstp27-package}
\alias{mgmtstp27}
\docType{package}
\title{
Additional tools for the analysis of methylation data
}
\description{
Additional tools for the analysis of methylation data
}
\details{
\tabular{ll}{
Package: \tab mgmtstp27\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-09-11\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Author: Pierre Bady <pierre.bady@unil.ch>\cr
Maintainer: Pierre Bady <pierre.bady@unil.ch>\cr
}
\references{
Bady, P., D. Sciuscio, A.-C. Diserens, J. Bloch, M. J. van den Bent, C. Marosi, P.-Y. Dietrich, M. Weller, L. Mariani, F. L. Heppner, D. R. McDonald, D. Lacombe, R. Stupp, M. Delorenzi, and M. E. Hegi. 2012. MGMT methylation analysis of glioblastoma on the Infinium methylation BeadChip identifies two distinct CpG regions associated with gene silencing and outcome, yielding a prediction model for comparisons across datasets, tumor grades, and CIMP-status. Acta Neuropathologica 124:547-560.\cr
}

